wip on auto publish and update
Working: if you make a draft release on github, `yarn release` will push to it.
Not working: if you don't make a draft 1st, it will say it is creating it but does not.

Probably working but have to test with a real install: autocheck for update
Will fail: it's not going to auto update without being signed